# storage/api-docs/api-docs.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/localhost"
        }
    ],
    "paths": {
...
        "\/api\/speech\/translate": {
            "post": {
                "tags": [
                    "Speech"
                ],
                "summary": "번역 결과 받기",
                "description": "번역할 텍스트를 보내면 번역 결과를 받을 수 있습니다. ko->ja",
                "parameters": [
                    {
                        "name": "AccessToken",
                        "in": "header",
                        "description": "Access Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {access_token}"
                    }
                ],
                "requestBody": {
                    "description": "Ko->ja 번역 텍스트 받기",
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/translateResult"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#\/components\/responses\/Success"
                    },
                    "400": {
                        "$ref": "#\/components\/responses\/BadRequest"
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/Unauthorized"
                    }
                }
            }
        },
...
    },
...
}